requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
anchorpy==0.8.1
ansi2html==1.7.0
anyio==3.4.0
apischema==0.16.6
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
backoff==1.11.1
base58==2.1.1
black==21.11b0
bleach==4.1.0
blinker==1.4
borsh-construct==0.1.0
Brotli==1.0.9
cachetools==4.2.4
ccxt==4.1.29
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorlog==6.6.0
construct==2.10.67
construct-typing==0.5.2
cryptography==36.0.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dnspython==2.2.0
driftpy==0.5.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.6.0
flake8==4.0.1
Flask==2.0.2
Flask-Compress==1.10.1
frozenlist==1.2.0
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonrpcclient==4.0.1
jsonrpcserver==5.0.6
jsons==1.6.1
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-dash==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
loguru==0.6.0
mango-explorer==3.4.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.12.0
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.8
nox==2022.1.7
nox-poetry==0.9.0
numpy==1.22.3
openpyxl==3.0.9
OSlash==0.6.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
plotly==5.3.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.23
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==7.0.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
Pygments==2.10.0
pyheck==0.1.4
Pympler==1.0.1
PyNaCl==1.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
pyserum==0.5.0a0
pytest==6.2.5
pytest-asyncio==0.17.2
pytest-xprocess==0.18.1
pythclient==0.0.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
retrying==1.3.3
rfc3986==1.5.0
Rx==3.2.0
rxpy-backpressure==1.0.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
solana==0.21.0
streamlit==1.7.0
sumtypes==0.1a6
tenacity==8.0.1
terminado==0.13.3
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.7.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
types-cachetools==4.2.6
types-requests==2.27.8
types-urllib3==1.26.9
typing-extensions==3.10.0.2
typish==1.9.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.13.1
visdcc==0.0.40
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.17.0
